feat(LIVE-32903): move analytics file
